Refactored the adding activity code to split common code into a common base class
